This commit rewrites the way pseudoranges are computed, now accounting for the receiver clock offset. It also adds more work in the QA code. If extra tests are activated by -DENABLE_UNIT_TESTING_EXTRA=ON or -DENABLE_SYSTEM_TESTING_EXTRA=ON, additional raw data files, a software-defined signal generator and GPSTk 2.9 are downloaded. Many fixes and code refactoring in tracking blocks.

+++ src/algorithms/telemetry_decoder/gnuradio_blocks/gps_l1_ca_telemetry_decoder_cc.h
@@ -121,8 +121,8 @@ private:
 
     double d_preamble_time_seconds;
 
-    double d_TOW_at_Preamble;
-    double d_TOW_at_current_symbol;
+    long double d_TOW_at_Preamble;
+    long double d_TOW_at_current_symbol;
 
     double Prn_timestamp_at_preamble_ms;
     bool flag_TOW_set;
